Move TaskValidator class
